Change all scss files to styled components and create the Navbar component
The navbar component is using hooks to control the menus.
