--- src/components/Navbar/navbar.scss
@@ -1,7 +1,0 @@
-.triangle {
-    width: 50px;
-    height: 50px;
-    border-left: 20px solid transparent;
-    border-right: 20px solid transparent;
-    border-top: 40px solid transparent;
-  }
